coletor-dados/requirements.txt
yfinance
pandas
pandas_ta
psycopg2-binary